test: split independent tests into separate files
